requirements.txt
# Frameworks e Servidores
fastapi
gunicorn
uvicorn

# Bibliotecas da Aplicação
google-generativeai
python-dotenv
Pillow

# pydantic é uma dependência do fastapi, mas é bom listá-lo
pydantic

redis